Move ewma calculation into its own file
